+++ R/GLMPoissonTest.R
@@ -11,6 +11,17 @@ GLMPoissonTest = function(sample.list, parameter) {
 
   if (call == FALSE | is.na(call)) {
 
+    # No parameters are defined
+    if (is.na(parameter[[2]])) {
+      larger = TRUE
+    }
+    else {
+      if (!all(names(parameter[[2]]) %in% c("larger"))) stop("Analysis model: GLMPoissonTest test: this function accepts only one argument (larger)")
+      # Parameters are defined but not the larger argument
+      if (!is.logical(parameter[[2]]$larger)) stop("Analysis model: GLMPoissonTest test: the larger argument must be logical (TRUE or FALSE).")
+      larger = parameter[[2]]$larger
+    }
+
     # Sample list is assumed to include two data frames that represent two analysis samples
 
     # Outcomes in Sample 1
@@ -28,11 +39,13 @@ GLMPoissonTest = function(sample.list, parameter) {
     colnames(data.complete) = c("TRT", "RESPONSE")
     data.complete$TRT=as.factor(data.complete$TRT)
 
-    # One-sided p-value
-    result = summary(stats::glm(RESPONSE ~ TRT, data = data.complete, family = poisson))$coefficients["TRT2", "Pr(>|z|)"]/2
+    # One-sided p-value (to be checked)
+    # result = summary(stats::glm(RESPONSE ~ TRT, data = data.complete, family = poisson))$coefficients["TRT2", "Pr(>|z|)"]/2
+    z = summary(stats::glm(RESPONSE ~ TRT, data = data.complete, family = poisson))$coefficients["TRT2", "z value"]
+    result = stats::pnorm(z, lower.tail = !larger)
+
   }
   else if (call == TRUE) {
-
     result=list("Poisson regression test")
   }