+++ R/PropTest.R
@@ -10,9 +10,28 @@ PropTest = function(sample.list, parameter) {
 
   if (call == FALSE | is.na(call)) {
 
-    # Yates' correction is set up by default to FALSE
-    if (is.na(parameter[[2]])) yates = FALSE
-    else yates = parameter[[2]]$yates
+    # No parameters are defined
+    if (is.na(parameter[[2]])) {
+      yates = FALSE
+      larger = TRUE
+    }
+    else {
+      if (!all(names(parameter[[2]]) %in% c("larger", "yates"))) stop("Analysis model: PropTest test: this function accepts only one argument (larger)")
+      # Yates' correction is set up by default to FALSE
+      if(is.null(parameter[[2]]$yates)) yates = FALSE
+      else {
+        if (!is.logical(parameter[[2]]$yates))
+          stop("Analysis model: PropTest test: the yates argument must be logical (TRUE or FALSE).")
+        yates = parameter[[2]]$yates
+      }
+      # Check if larger treatment effect is expected for the second sample or not (default = TRUE)
+      if (is.null(parameter[[2]]$larger)) larger = TRUE
+      else {
+        if (!is.logical(parameter[[2]]$larger))
+          stop("Analysis model: PropTest test: the larger argument must be logical (TRUE or FALSE).")
+        larger = parameter[[2]]$larger
+      }
+    }
 
     # Sample list is assumed to include two data frames that represent two analysis samples
 
@@ -27,11 +46,13 @@ PropTest = function(sample.list, parameter) {
     outcome2.complete = outcome2[stats::complete.cases(outcome2)]
 
     # One-sided p-value (treatment effect in sample 2 is expected to be greater than in sample 1)
-    result = stats::prop.test(c(sum(outcome2.complete), sum(outcome1.complete)),
-                              n = c(length(outcome2.complete), length(outcome1.complete)), alternative = "greater", correct = yates)$p.value
+    if (larger) result = stats::prop.test(c(sum(outcome2.complete), sum(outcome1.complete)),
+                                          n = c(length(outcome2.complete), length(outcome1.complete)), alternative = "greater", correct = yates)$p.value
+    else result = stats::prop.test(c(sum(outcome2.complete), sum(outcome1.complete)),
+                                   n = c(length(outcome2.complete), length(outcome1.complete)), alternative = "less", correct = yates)$p.value
+
   }
   else if (call == TRUE) {
-
     result=list("Test for proportions")
   }