+++ R/PropTestNI.R
@@ -20,8 +20,19 @@ PropTestNI = function(sample.list, parameter) {
     margin = as.numeric(parameter[[2]]$margin)
 
     # Yates' correction is set up by default to FALSE
-    if (is.null(parameter[[2]]$yates)) yates = FALSE
-    else yates = parameter[[2]]$yates
+    if(is.null(parameter[[2]]$yates)) yates = FALSE
+    else {
+      if (!is.logical(parameter[[2]]$yates))
+        stop("Analysis model: PropTestNI test: the yates argument must be logical (TRUE or FALSE).")
+      yates = parameter[[2]]$yates
+    }
+    # Check if larger treatment effect is expected for the second sample or not (default = TRUE)
+    if (is.null(parameter[[2]]$larger)) larger = TRUE
+    else {
+      if (!is.logical(parameter[[2]]$larger))
+        stop("Analysis model: PropTestNI test: the larger argument must be logical (TRUE or FALSE).")
+      larger = parameter[[2]]$larger
+    }
 
     # Sample list is assumed to include two data frames that represent two analysis samples
 
@@ -36,15 +47,17 @@ PropTestNI = function(sample.list, parameter) {
     outcome2.complete = outcome2[stats::complete.cases(outcome2)]
 
     # One-sided p-value (treatment effect in sample 2 is expected to be greater than in sample 1)
-    result = stats::prop.test(c(sum(outcome2.complete) + margin * length(outcome2.complete),
-                                sum(outcome1.complete)),
-                              n = c(length(outcome2.complete), length(outcome1.complete)),
-                              alternative = "greater", correct = yates)$p.value
+    if (larger) result = stats::prop.test(c(min(sum(outcome2.complete) + margin * length(outcome2.complete), length(outcome2.complete)),
+                                            sum(outcome1.complete)),
+                                          n = c(length(outcome2.complete), length(outcome1.complete)),
+                                          alternative = "greater", correct = yates)$p.value
+    else result = stats::prop.test(c(min(sum(outcome1.complete) + margin * length(outcome1.complete), length(outcome1.complete)),
+                                     sum(outcome2.complete)),
+                                   n = c(length(outcome1.complete), length(outcome2.complete)),
+                                   alternative = "greater", correct = yates)$p.value
   }
   else if (call == TRUE) {
-
-    result=list("Test for proportions", "Non-inferiority margin = ")
-
+    result=list("Non-inferiority test for proportions")
   }
 
   return(result)