+++ R/TTest.R
@@ -11,6 +11,18 @@ TTest = function(sample.list, parameter) {
 
   if (call == FALSE | is.na(call)) {
 
+    # No parameters are defined
+    if (is.na(parameter[[2]])) {
+      larger = TRUE
+    }
+    else {
+      # Check the name of arguments
+      if (!all(names(parameter[[2]]) %in% c("larger"))) stop("Analysis model: Ttest test: this function accepts only one argument (larger)")
+      # larger argument
+      if (!is.logical(parameter[[2]]$larger)) stop("Analysis model: TTest test: the larger argument must be logical (TRUE or FALSE).")
+      larger = parameter[[2]]$larger
+    }
+
     # Sample list is assumed to include two data frames that represent two analysis samples
 
     # Outcomes in Sample 1
@@ -24,7 +36,9 @@ TTest = function(sample.list, parameter) {
     outcome2.complete = outcome2[stats::complete.cases(outcome2)]
 
     # One-sided p-value (treatment effect in sample 2 is expected to be greater than in sample 1)
-    result = stats::t.test(outcome2.complete, outcome1.complete, alternative = "greater")$p.value
+    if (larger) result = stats::t.test(outcome2.complete, outcome1.complete, alternative = "greater")$p.value
+    else result = stats::t.test(outcome2.complete, outcome1.complete, alternative = "less")$p.value
+
   }
   else if (call == TRUE) {
     result=list("Student's t-test")